+++ wasm-crypto-wasm/bundle.ts
@@ -1,7 +1,7 @@
 // Copyright 2019-2022 @polkadot/wasm-crypto-wasm authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { base64Decode, unzlibSync } from 'https://deno.land/x/polkadot@0.2.18/wasm-util/mod.ts';
+import { base64Decode, unzlibSync } from 'https://deno.land/x/polkadot/wasm-util/mod.ts';
 
 import { bytes, lenIn, lenOut } from './deno/bytes.js';
 
@@ -16,4 +16,4 @@ export { packageInfo } from './packageInfo.ts';
  * Upon loading the exposed bytes will be decoded and decompressed form this
  * specific format and returned.
  */
-export const wasmBytes = unzlibSync(base64Decode(bytes, new Uint8Array(lenIn)), new Uint8Array(lenOut));
+export const wasmBytes = /*#__PURE__*/ unzlibSync(base64Decode(bytes, new Uint8Array(lenIn)), new Uint8Array(lenOut));